Refactor health data upload into HealthRepository
This extracts the duplicate upload batching and HTTP post loop out
of UploadToShelfService and moves it to HealthRepositoryImpl.
UploadToShelfService is now a thin coordinator that just calls the
repository. All tests updated successfully.
