+++ app/src/main/java/org/ole/planet/myplanet/services/UploadToShelfService.kt
@@ -7,9 +7,9 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 import javax.inject.Singleton
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.supervisorScope
 import kotlinx.coroutines.sync.Semaphore
 import kotlinx.coroutines.sync.withPermit
@@ -19,7 +19,6 @@ import org.ole.planet.myplanet.data.DatabaseService
 import org.ole.planet.myplanet.data.api.ApiInterface
 import org.ole.planet.myplanet.di.AppPreferences
 import org.ole.planet.myplanet.di.ApplicationScope
-import org.ole.planet.myplanet.model.RealmHealthExamination.Companion.serialize
 import org.ole.planet.myplanet.repository.HealthRepository
 import org.ole.planet.myplanet.repository.UserRepository
 import org.ole.planet.myplanet.repository.UserSyncRepository
@@ -110,34 +109,7 @@ class UploadToShelfService @Inject constructor(
     fun uploadHealth() {
         appScope.launch(dispatcherProvider.io) {
             val myHealths = healthRepository.getUpdatedHealthExaminations()
-
-            val uploadedHealths = mutableMapOf<String, String?>()
-            val semaphore = Semaphore(5)
-            supervisorScope {
-                myHealths.map { pojo ->
-                    async {
-                        semaphore.withPermit {
-                            try {
-                                val res = apiInterface.postDoc(UrlUtils.header, "application/json", "${UrlUtils.getUrl()}/health", serialize(pojo))
-
-                                if (res.body() != null && res.body()?.has("id") == true) {
-                                    val rev = res.body()?.get("rev")?.asString
-                                    val id = pojo._id
-                                    if (id != null) {
-                                        return@async id to rev
-                                    }
-                                }
-                            } catch (e: Throwable) {
-                                e.printStackTrace()
-                            }
-                            null
-                        }
-                    }
-                }.awaitAll().filterNotNull().forEach { (id, rev) ->
-                    uploadedHealths[id] = rev
-                }
-            }
-
+            val uploadedHealths = healthRepository.uploadHealthData(myHealths)
             healthRepository.markHealthExaminationsUploaded(uploadedHealths)
         }
     }
@@ -148,39 +120,7 @@ class UploadToShelfService @Inject constructor(
                 if (userId.isNullOrEmpty()) return@launch
 
                 val myHealths = healthRepository.getUpdatedHealthForUser(userId)
-
-                val uploadedHealths = mutableMapOf<String, String?>()
-                val semaphore = Semaphore(5)
-                supervisorScope {
-                    myHealths.map { pojo ->
-                        async {
-                            semaphore.withPermit {
-                                try {
-                                    val res = apiInterface.postDoc(
-                                        UrlUtils.header,
-                                        "application/json",
-                                        "${UrlUtils.getUrl()}/health",
-                                        serialize(pojo)
-                                    )
-
-                                    if (res.body() != null && res.body()?.has("id") == true) {
-                                        val rev = res.body()?.get("rev")?.asString
-                                        val id = pojo._id
-                                        if (id != null) {
-                                            return@async id to rev
-                                        }
-                                    }
-                                } catch (e: Throwable) {
-                                    e.printStackTrace()
-                                }
-                                null
-                            }
-                        }
-                    }.awaitAll().filterNotNull().forEach { (id, rev) ->
-                        uploadedHealths[id] = rev
-                    }
-                }
-
+                val uploadedHealths = healthRepository.uploadHealthData(myHealths)
                 healthRepository.markHealthExaminationsUploaded(uploadedHealths)
 
                 withContext(dispatcherProvider.main) {
